fix: timeAsDouble doesn't exist (newer versions of Mirage require it)
Later versions of Mirage make use of the timeAsDouble property that our FakeUnityEngine doesn't have. This is a potentially dirty hack. Previously we used a float, but now we're using a double, and if we reference Time.time from FakeUnityEngine, we'll cast down the double to a single (float). timeAsDouble was also introduced in newer Unity versions for better precision, I believe?